pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robust-autonomy-stack"
version = "0.1.0"
description = "Self-healing autonomous driving stack with learned risk monitoring and RL red-teaming"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "numpy==1.24.2",
    "scipy>=1.10.0",
    "opencv-python>=4.8.0",
    "shapely>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "stable-baselines3==2.2.1",
    "gymnasium==0.28.1",
    "metadrive-simulator==0.4.2.2",
    "torch>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
]

[project.scripts]
robust-autonomy-stack = "robust_autonomy_stack.cli:main"

[tool.setuptools]
packages = ["robust_autonomy_stack"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
